Refactoring: Post2dWindowNodeVectorParticleGroupStructuredDataItem::SetSettingCommand is moved to separate files.
